Finished symmetry type determination
The code can now read all the information from the output of findsym
It then determines how each symmetry transforms each sulattice
This information is then used by symmetr in symmetrize_sympy.py to determine if
symmetry should be considered or not.

It will also output the resulting tensors in both the transformed basis used for symmetries and
the original basis.

!!!Not tested!!! Most likely it contains bugs, possibly even severe. But it runs and produces result, which look reasonable.

It remains to be done:
* Renaming of the tensor transformed to the original basis
* Hexagonal symmetry operations
* Comments!