Added hexagonal symmetry operations
Symmetry operations can now be used as well, however this is not tested at all!
It will most likely contain bugs if hexagonal symmetry operations are present,
however it seems to work fine for the case of Mn2Au.
The algorithm used is quite primitive, it should give correct results, but the matrices it outputs may not be the most general ones.